safari-600.7-branch/LayoutTests/platform/ios-sim/http/tests/security/contentSecurityPolicy/object-src-none-blocked-expected.txt

CONSOLE MESSAGE: Refused to load plugin data from 'data:application/x-webkit-test-netscape,logifloaded' because it violates the following Content Security Policy directive: "object-src 'none'".

CONSOLE MESSAGE: Refused to load plugin data from 'data:application/x-webkit-test-netscape,logifloaded' because it violates the following Content Security Policy directive: "object-src 'none'".



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
